+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
   - Should close activity when failed to load annotated verses, reading progress, and Strong's numbers
   - Properly nullify when jobs updating annotated verses complete
   - Should scroll after chapter group is expanded / collapsed
+  - Fixed issue with Direct Share on Android 10 and above
 - Changes:
   - No longer show book name when simple reading mode is off
 - Refactoring: